requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
certifi==2018.8.24
chardet==3.0.4
discord.py==0.16.12
idna==2.7
multidict==4.4.1
oauthlib==2.1.0
pkg-resources==0.0.0
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
websockets==3.4
